requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
openai>=1.0.0
apscheduler
pygame
smtplib
email
shelve
pydantic